don't save huge cookies
Huge cookies confuse the code that reads them back in. I suppose there's
always the possibility of getting a cookies.txt from something else, or
a manually-edited one, but...

Incidentally, the RFC thinks user agents should allow 4096 bytes for
name+value+attributes, which we must be within, say, 50 bytes of.