don't reuse user- and useragent-stylesheet
The notMatchingBefore variable in CssSelector is modified as a page is
styled. Thererfore we can't share a single copy of the user- and
useragent-stylesheets between CssContexts.

Testcase:

<html>
<body>
<a href=whatever.html>
 <img alt=test src=whatever.jpg>
</a>
<div>
  <img src=something.jpg>
</div>
</body>
</html>

On reload the first image looses it's border.

Tests with gettimeofday showed times below 1ms to create user- and
useragent-styles on a Pentium-M 1.8 GHz laptop.

reported-by: corvid <corvid@lavabit.com>