digest auth: in Authorization header, quote digest-uri-value
found by furaisanjin.
http://www.rfc-editor.org/errata_search.php?rfc=2617 has an entry
showing that this should be done. Now, the RFC was done in 1999, and
the error was reported in 2010 and verified in 2011, so we're lucky
with our timing, but isn't it surprising how slow of a process it is
to clean out the corners like this in important specifications?