can't use fl_oldfocus anymore -- this will make focus behaviour somewhat screwy
http://lists.dillo.org/pipermail/dillo-dev/2014-November/010273.html

A couple of months ago, Lauri Kasanen made some changes to fltk's linker
stuff, and I guess we're seeing the effect here.

As for focus behaviour being screwy, if you click on a page to give it
focus, then move the cursor away and then back onto the page, it won't
get focus again -- focus will go to an input or the location bar or
something.